Fix comment about trace() function (#194)
The comment stated that `pow == true` is about the number of steps,
but the implementation was making the number of blocks to be a power
of two. These are different especially when `k` is not a power-of-two.

This commit fixes the comment according to the
implementation. Otherwise, the operation might not terminate (for
example when `k` is three).