Add and fix smoke tests on prove and verify (#214)
* Add lightweight examples

This commit adds two examples that fail on
c470d637134797162546321140dbb213dd1761a2

The examples are kept small so that they can be cheaply tested in
the CI.

* Add smoke tests in the CI

The new tests mimick what's written in README.md. This is meant as
a smoke test to see if any example works at all. That was not the
case on the main branch after c470d637134797162546321140dbb213dd1761a2.

* Fix failing smoke test

block.regs was immutable throughout the k instructions, so it was
not an appropriate thing to parse here.

* Use the constant one more time

* De-dup cleanup() in smoke.sh

* Finish incomplete renaming

* Move only when source exists

Otherwise, ctrl-c during `cargo build` caused an error at the `mv`
invocation in `cleanup()`. In this case, `cleanup()` was called twice
as an error handler and the ctrl-c handler.