refactor: update column name map to btree (#63)
Prior to this commit, the column name map had an underlying hash map
that would make a column name to a range within the constraint system.

This map had to be duplicated into a vector, so its order could be
preserved for iteration purposes.

This commit updates the underlying map to a btree map, that via its
properties retains the order of the columns.

The column set is defined statically via an iterator of the enum that
declares the columns. This commit removes the manual allocation of the
map (previously via `Self::allocate_bulk`), and replaces it with static
allocation.